+++ crates/compiler/src/lib.rs
@@ -3,6 +3,7 @@ use parser::{RawExpr, RawStmt};
 use crate::bytecode::{AstLoc, Instr};
 
 pub mod bytecode;
+mod utils;
 
 // pub fn compile(prog: parser::Program) -> Code {
 //     let mut fn_map = SlotMap::with_key();
@@ -39,14 +40,6 @@ struct Local<'s> {
     depth: i32,
 }
 
-#[must_use]
-#[derive(Debug)]
-struct JumpHelper(usize);
-
-#[must_use]
-#[derive(Debug)]
-struct BackJumpHelper(usize);
-
 impl<'a, 's> FnComping<'a, 's> {
     fn add_block(&mut self, block: &'a parser::Block<'s>) {
         self.begin_scope();
@@ -77,47 +70,6 @@ impl<'a, 's> FnComping<'a, 's> {
 
     fn n_instrs(&self) -> usize {
         self.output.code.len()
-    }
-
-    fn add_loop(
-        &mut self,
-        ty: impl Fn(usize) -> Instr<'static>,
-        stmt: &'a parser::Stmt<'s>,
-    ) -> JumpHelper {
-        self.add_instr_sloc(ty(0), stmt);
-        JumpHelper(self.n_instrs())
-    }
-
-    fn add_jump_back_to_point(&self) -> BackJumpHelper {
-        BackJumpHelper(self.output.code.len())
-    }
-
-    fn patch_fjump(&mut self, h: JumpHelper) {
-        let current_ip = self.n_instrs();
-        let diff = current_ip - h.0;
-
-        if let Instr::JumpForward(i) | Instr::JumpForwardIfFalse(i) = &mut self.output.code[h.0 - 1]
-        {
-            *i = diff;
-        } else {
-            unreachable!("{:?}, {:?}", self.output.code, h);
-        }
-    }
-
-    /// Add [`Instr::JumpForwardIfFalse`]
-    fn add_jfif(&mut self, stmt: &'a parser::Stmt<'s>) -> JumpHelper {
-        self.add_loop(Instr::JumpForwardIfFalse, stmt)
-    }
-
-    /// Add [`Instr::JumpForward`]
-    fn add_jf(&mut self, stmt: &'a parser::Stmt<'s>) -> JumpHelper {
-        self.add_loop(Instr::JumpForward, stmt)
-    }
-
-    fn add_jb(&mut self, pos: BackJumpHelper, stmt: &'a parser::Stmt<'s>) {
-        let ip = self.output.code.len();
-        let diff = ip - pos.0;
-        self.add_instr_sloc(Instr::JumpBackward(diff), stmt)
     }
 
     pub fn push_stmt(&mut self, stmt: &'a parser::Stmt<'s>) {
